requirements.txt
streamlit
google-generativeai>=0.7.0
weasyprint